Fix sensors requiring reboot issue [UNVERIFIED]
[UNVERIFIED FIX]

Attempt to address issue when sensors stop responding and reboot is required. Methodology: daveenguyen suspected worker thread is being killed. After inspecting code, and the log messages, it seems like an OSError would cause this. 

(1) Attempt removing the break statement to give worker thread a chance to continue running if this is indeed where it was getting killed. Add log info to detect if this occurs.
(2) Add additional except: clause for any non OSError exception in case this is what is causing worker thread to die. Add log info to detect if this occurs.

This will require testing to verify whether or not the fix was successful. Success criteria: logs indicate that either (1) or (2) successfully saved the worker thread from being killed, and sensors are still responsive. Failure criteria: Sensors become unresponsive.

Publishing this attempt, so if others also want to try, they are able, since reproducing the issue may take days to weeks to occur for me.